+++ examples/serversession-example-yesod-persistent/Application.hs
@@ -6,6 +6,7 @@ module Application
     , appMain
     , develMain
     , makeFoundation
+    , makeLogWare
     -- * for DevelMain
     , getApplicationRepl
     , shutdownApp
@@ -19,6 +20,7 @@ import Database.Persist.Sqlite              (createSqlitePool, runSqlPool,
                                              sqlDatabase, sqlPoolSize)
 import Import
 import Language.Haskell.TH.Syntax           (qLocation)
+import Network.Wai (Middleware)
 import Network.Wai.Handler.Warp             (Settings, defaultSettings,
                                              defaultShouldDisplayException,
                                              runSettings, setHost,
@@ -30,9 +32,7 @@ import Network.Wai.Middleware.RequestLogger (Destination (Logger),
 import System.Log.FastLogger                (defaultBufSize, newStdoutLoggerSet,
                                              toLogStr)
 
-import qualified Data.Proxy as P
-import qualified Web.ServerSession.Core as SS
-import qualified Web.ServerSession.Backend.Persistent as SS
+import Web.ServerSession.Backend.Persistent
 
 -- Import all relevant handler modules here.
 -- Don't forget to add new modules to your cabal file!
@@ -47,7 +47,7 @@ mkYesodDispatch "App" resourcesApp
 
 -- Create migration function using both our entities and
 -- serversession-backend-persistent ones.
-mkMigrate "migrateAll" (SS.serverSessionDefs (P.Proxy :: P.Proxy SS.SessionMap) ++ entityDefs)
+mkMigrate "migrateAll" (entityDefs `embedEntityDefs` serverSessionDefsBySessionMap)
 
 
 -- | This function allocates resources (such as a database connection pool),
@@ -85,10 +85,17 @@ makeFoundation appSettings = do
     return $ mkFoundation pool
 
 -- | Convert our foundation to a WAI Application by calling @toWaiAppPlain@ and
--- applyng some additional middlewares.
+-- applying some additional middlewares.
 makeApplication :: App -> IO Application
 makeApplication foundation = do
-    logWare <- mkRequestLogger def
+    logWare <- makeLogWare foundation
+    -- Create the WAI application and apply middlewares
+    appPlain <- toWaiAppPlain foundation
+    return $ logWare $ defaultMiddlewaresNoLogging appPlain
+
+makeLogWare :: App -> IO Middleware
+makeLogWare foundation =
+    mkRequestLogger def
         { outputFormat =
             if appDetailedRequestLogging $ appSettings foundation
                 then Detailed True
@@ -98,10 +105,6 @@ makeApplication foundation = do
                             else FromSocket)
         , destination = Logger $ loggerSet $ appLogger foundation
         }
-
-    -- Create the WAI application and apply middlewares
-    appPlain <- toWaiAppPlain foundation
-    return $ logWare $ defaultMiddlewaresNoLogging appPlain
 
 -- | Warp settings for the given foundation value.
 warpSettings :: App -> Settings