+++ serversession-backend-persistent/tests/Main.hs
@@ -3,18 +3,16 @@ module Main (main) where
 import Control.Monad (forM_)
 import Control.Monad.Logger (runStderrLoggingT, runNoLoggingT)
 import Data.Pool (destroyAllResources)
-import Data.Proxy (Proxy(..))
 import Database.Persist.Postgresql (createPostgresqlPool)
 import Database.Persist.Sqlite (createSqlitePool)
 import Test.Hspec
 import Web.ServerSession.Backend.Persistent
-import Web.ServerSession.Core (SessionMap)
 import Web.ServerSession.Core.StorageTests
 
 import qualified Database.Persist.TH as P
 import qualified Database.Persist.Sql as P
 
-P.mkMigrate "migrateAll" (serverSessionDefs (Proxy :: Proxy SessionMap))
+P.mkMigrate "migrateAll" serverSessionDefsBySessionMap
 
 main :: IO ()
 main = hspec $